Restore serialization of Status
In 2.0-M6-SNAP6, Status events could be serialized without issue. I
suspect the problem is caused by the addition of CountDownLatch to
CompositeStatus in 6e17adfebd1300bfce8880155ddb2c67ed5d3a21. This
patch marks the latch transient in StatefulStatus,
ScalaTestStatefulStatus, and CompositeStatus.

Further, StatefulStatusSpec was in the src/main directory and as a
result, wasn't getting run during "ant test". I assume this was a
mistake so I moved it to src/test.