Fix isNestedMethod() detection to include a corner case
Nested methods can end with a number including 0. For example,
consider a block like

    class ExampleSpec extends Spec {
      object `An example` {
          def `test 1` {
            def helper() = ...
          }

          def `test 2` {
            def helper() = ...
          }

          ...

          def `test 10` {
            def helper() = ...
          }
        }
      }

The `helper` methods will be turned into something like:

    `ExampleSpec$An$u0020example$$helper$1`
    `ExampleSpec$An$u0020example$$helper$2`
    ...
    `ExampleSpec$An$u0020example$$helper$10`

But that tenth helper method will currently be counted as a test and executed.
(This extra execution is most obvious when the helper method throws an exception.)