#1519 (checkpoint): test improvements for CommonGenerators
Started out by refactoring some of the ubiquitous boilerplate into a common method, so that many of these tests can become one-liners. Have only applied that to a couple of the tests, but in principle should be applied much more widely.

In spot-checking that, discovered that samplesForGen() is and always has been broken, such that it was never actually testing anything, and many tests have been mostly no-ops.

Fixing *that* showed that the tests for function1/2s have always been fundamentally broken -- they're testing something that can't possibly work, because the functions don't have equality defined in a useful way. After discussion with Bill, replaced the broken tests with testing the output of those functions, rather than their identity.