Improve startup performance by 80ms
Scala reflection is very slow because it takes a long time to load all
of the classes needed by the runtime mirror. Java reflection, on the
other hand, is reasonably quick. In unscientific and ad-hoc benchmarkes,
I found that the startup time decreased by about 80ms with this change.
Given java reflection was already used for scala 2.10, the risk of this
change seems minimal.