Added poll method to ConcurrentLinkedQueue wrappers. Fixed two race conditions in InternalFutureOutcome. The one that was causing ConcurrentModificationExeption was that we registered callback on the future passed to InternalFutureOutcome's in its constructor, so that could fire before the execution of the callback (which sends the test completion event to the reporter) that AsyncEngine next registered on the InternalFutureOutcome via its onComplete method. In that case, the Informer was getting swapped out before the currently running test had sent its test completion event. Fixed this by moving the future.onComplete call out of the InternalFutureOutcome method into the InternalFutureOutcomes's onComplete method itself. This means the InternalFutureOutcome will "hang" if no one ever calls onComplete, but it is an internal class and AsyncEngine always calls onComplete.
Also fixed a race condition that was not related to the ConcurrentModificationBug we are fixing,
which was essentially because we were using the iterator method on Java's ConcurrentLinkedQueue,
but treating it as if it would pick up any new element added while were were iterating. But the
Javadoc says the iterator may not return such newly added elements. Still need to close this
loophole in Status, but the ConcurrentModificationException when using Async styles with
a thread-pool execution context bug appears to be fixed.