Disable some tests for dotty
`CompileMacro.{assertCompileImpl, mustNotCompileImpl}` are currently impossible to implement.
The issue is that the macros are defined using extention methods in the Scala2 by creating an implicit conversion. The macro implemetation is there to strip the artefacts of the implicit conversion.

Those kinds of extention methods will not be suported in Dotty and will need to migrate to proper extension methods. In that scenario those macros will become redundant and we would just call typeChecks directly.