Avoid polling Scala futures
Futures contained a futureValueImpl which would poll the future.

Futures is implemented by ScalaFutures and JavaFutures, where JavaFutures had an
override for this method to avoid the polling.

This PR moves the futureValueImpl implementation to ScalaFutures (since it was only
used there anyway), and removes the need to poll.

This also removes (the need for) FuturesSpec, since that mainly tested the polling
mechanism that is now neither used in the Java implementation nor in the Scala one.
The actual logic is tested in JavaFuturesSpec and ScalaFuturesSpec respectively.