be Scala 2.13.3 friendly (Symbol#toString changed)
the old code relied on this assumption:
`mySymbol.name == mySymbol.toString.drop(1)`
but Scala 2.13.3 is changing Symbol#toString to be
e.g. `Symbol(foo)` instead of `'foo`

the new code works fine on any Scala version.  I cannot see
any reason for the old code to be the way it was