Fix #2116: Use an object identity instead of UUID in TestSortingReporter.
The UUIDs are only used as an in-memory identifier for `Slot`s.
It is more efficient and more reliable to use an object identity
instead, for that purpose.

This removes the dependency on `UUID.randomUUID()`, which unblocks
using ScalaTest with Scala.js >= 1.10.0. See
https://github.com/scala-js/scala-js/security/advisories/GHSA-j2f9-w8wh-9ww4
and
https://github.com/scala-js/scala-js/issues/4657
for the reason why `randomUUID()` fails to link by default in
recent Scala.js.