feat: set sorting reporter as complete if test run failed unexpectedly
If test runner is of sorted type, it buffers all events and eventually emits/prints them upon completion of the test. Now, in case test runner fails unexpectedly it does not flush buffered events until an internal timer kicks off; this won't work with SBT system as all loggers of a task are discarded by the time that internal timer is triggered.